SplitMergeFunction tests and impl fixes.
